v2/requirements.txt
bs4
discord.py[voice]
httpx
youtube-dl
